requirements.txt
asgiref
certifi
charset-normalizer
Django
django-environ
idna
psycopg2-binary
requests
setuptools
sqlparse
tzdata
urllib3
